+++ src/main/java/com/sparta/gitandrun/order/repository/OrderRepository.java
@@ -1,7 +1,7 @@
 package com.sparta.gitandrun.order.repository;
 
 import com.sparta.gitandrun.order.entity.Order;
-import com.sparta.gitandrun.order.repository.queryDsl.OrderCustomRepository;
+import com.sparta.gitandrun.order.repository.custom.OrderCustomRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
